Revert "automatische schaling en auto formatting"
This reverts commit adc3e8cd8f5d39297638b9865cd0ca3a23c3aa22.

# Conflicts:
#	src/main/java/org/Game/PowerTimer.java
#	src/main/java/org/Game/Scenes/GameScene.java
#	src/main/java/org/Game/SkySourerGame.java

+++ src/main/java/org/Game/Entities/Buttons/ButtonSpeed.java
@@ -2,23 +2,24 @@ package org.Game.Entities.Buttons;
 
 import com.github.hanyaeger.api.Coordinate2D;
 import org.Game.Scenes.StartScene;
+import org.Game.SkySourerGame;
 import javafx.scene.input.MouseButton;
 
 
 public class ButtonSpeed extends Button {
 
-    private StartScene startScene;
-    private double speed;
+	private StartScene startScene;
+	private int speed;
 
-    public ButtonSpeed(int width, int height, Coordinate2D location, String text, StartScene startScene, double speed) {
-        super(width, height, location, text);
-        this.startScene = startScene;
-        this.speed = speed;
-    }
+	public ButtonSpeed(int width, int height, Coordinate2D location, String text, StartScene startScene, int speed) {
+		super(width, height, location, text);
+		this.startScene = startScene;
+		this.speed = speed;
+	}
 
 
-    @Override
-    public void onMouseButtonPressed(MouseButton button, Coordinate2D coordinate2D) {
-        startScene.setSpeed(speed);
-    }
+	@Override
+	public void onMouseButtonPressed(MouseButton button, Coordinate2D coordinate2D) {
+		startScene.setSpeed(speed);
+	}
 }
