+++ src/main/java/org/Game/Entities/Power.java
@@ -3,9 +3,13 @@ package org.Game.Entities;
 
 import com.github.hanyaeger.api.Coordinate2D;
 import com.github.hanyaeger.api.entities.Collided;
+import com.github.hanyaeger.api.entities.Collider;
 import com.github.hanyaeger.api.entities.SceneBorderCrossingWatcher;
 import com.github.hanyaeger.api.entities.impl.DynamicSpriteEntity;
+import com.github.hanyaeger.api.scenes.SceneBorder;
 import org.Game.Scenes.GameScene;
+
+import java.util.List;
 
 public abstract class Power extends DynamicSpriteEntity implements Collided, SceneBorderCrossingWatcher {
 	protected int hight;
@@ -14,7 +18,7 @@ public abstract class Power extends DynamicSpriteEntity implements Collided, Sce
 
 	protected GameScene gameScene;
 
-    public Power(Coordinate2D location, double gameSpeed, GameScene gameScene, String filename) {
+    public Power(Coordinate2D location, int gameSpeed, GameScene gameScene, String filename) {
         super(filename, location);
     }
 }
